chore: e2e tsx 도입과 mock 서버 TS 전환
