+++ package.json
@@ -14,7 +14,9 @@
     "test:e2e": "playwright test",
     "test:e2e:ui": "playwright test --ui",
     "test:e2e:debug": "playwright test --debug",
-    "e2e:mock": "node e2e/mock-server/server.mjs",
+    "test:e2e:update": "playwright test --update-snapshots",
+    "test:e2e:type-check": "tsc -p e2e/tsconfig.json --noEmit",
+    "e2e:mock": "tsx e2e/mock-server/server.ts",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
     "storybook": "pnpm --filter @muneo/design-system storybook",
@@ -30,6 +32,7 @@
     "husky": "^9.1.7",
     "lint-staged": "^16.4.0",
     "prettier": "^3.8.1",
+    "tsx": "^4.21.0",
     "turbo": "^2.8.20",
     "typescript": "^5.9.3"
   },
